requirements.txt
langchain
langchain-openai
langchain-community
langgraph
openai
python-dotenv
pydantic
streamlit
reportlab
duckduckgo-search
